refactor: extract useNativeYieldVault into shared logic
