+++ packages/suite/src/views/wallet/tokens/common/TokensTable/hooks/useTokenYieldBadge.ts
@@ -1,5 +1,6 @@
 import { useMemo } from 'react';
 
+import { selectBestEnabledYieldVault } from '@suite-common/earn-stablecoin';
 import { type YieldDtoV2 } from '@suite-common/earn-stablecoin-api';
 import { type EnhancedTokenInfo } from '@suite-common/token-definitions';
 import { type NetworkSymbol } from '@suite-common/wallet-config';
@@ -12,7 +13,6 @@ import { getApyPercent } from '@suite-common/wallet-utils';
 import { type TokenInfo } from '@trezor/blockchain-link-types';
 import { exhaustive } from '@trezor/type-utils';
 
-import { getBestEnabledYieldVault } from 'src/components/earn/utils/yieldVaultUtils';
 import { useSelector } from 'src/hooks/suite';
 
 import type { TokensTableType } from '../types';
@@ -83,7 +83,7 @@ export const useTokenYieldBadge = ({
     // A vault the user already deposited into states the rate they actually earn, so it
     // outranks a higher-paying one they have not touched.
     const bestEnabledVault = useSelector(state =>
-        getBestEnabledYieldVault(
+        selectBestEnabledYieldVault(
             state,
             vaultsWithPosition.length > 0 ? vaultsWithPosition : matchedVaults,
         ),
