+++ suite-common/earn-stablecoin/src/index.ts
@@ -1,3 +1,4 @@
 export * from './allowance';
 export * from './signing';
 export * from './tx-simulation';
+export * from './vaults';
