+++ suite-common/earn-stablecoin/src/vaults/getWrappedNativeYieldVaults.test.ts
