+++ suite-common/earn-stablecoin/src/vaults/getWrappedNativeYieldVaults.ts
