+++ suite-common/earn-stablecoin/src/vaults/index.ts
@@ -1,0 +1,3 @@
+export * from './getWrappedNativeYieldVaults';
+export * from './useNativeYieldVault';
+export * from './yieldVaultSelectors';
