+++ suite-common/earn-stablecoin/tsconfig.json
@@ -4,8 +4,10 @@
     "references": [
         { "path": "../calldata" },
         { "path": "../earn-stablecoin-api" },
+        { "path": "../message-system" },
         { "path": "../schemas" },
         { "path": "../wallet-config" },
+        { "path": "../wallet-core" },
         { "path": "../wallet-types" },
         { "path": "../wallet-utils" },
         {
