+++ suite-native/banners/package.json
@@ -16,6 +16,7 @@
         "@suite-common/analytics": "workspace:*",
         "@suite-common/dependency-injection": "workspace:*",
         "@suite-common/device": "workspace:*",
+        "@suite-common/earn-stablecoin": "workspace:*",
         "@suite-common/message-system": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
         "@suite-common/wallet-types": "workspace:*",
