+++ suite-native/module-accounts-management/tsconfig.json
@@ -8,6 +8,9 @@
         },
         { "path": "../../suite-common/device" },
         {
+            "path": "../../suite-common/earn-stablecoin"
+        },
+        {
             "path": "../../suite-common/earn-stablecoin-api"
         },
         {
